requirements.txt
Flask>=2.0.1
Flask-JWT-Extended==4.3.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.1
marshmallow==3.14.1
flask-swagger-ui==3.36.0
flask-restful-swagger-2==0.35
flask-cors==3.0.10
PyJWT==2.5.0
